+++ middleman-core/lib/middleman-core/sitemap/resource.rb
@@ -1,5 +1,6 @@
+require 'rack'
 require 'middleman-core/sitemap/extensions/traversal'
-require 'middleman-core/sitemap/extensions/content_type'
+require 'middleman-core/sitemap/extensions/metadata'
 require 'middleman-core/file_renderer'
 require 'middleman-core/template_renderer'
 
@@ -9,11 +10,13 @@ module Middleman
     # Sitemap Resource class
     class Resource
       include Middleman::Sitemap::Extensions::Traversal
-      include Middleman::Sitemap::Extensions::ContentType
 
       # @return [Middleman::Application]
       attr_reader :app
       delegate :logger, :instrument, to: :app
+
+      attr_reader :metadata
+      delegate :data, :raw_data, to: :metadata
 
       # @return [Middleman::Sitemap::Store]
       attr_reader :store
@@ -27,12 +30,16 @@ module Middleman
       # @return [String]
       attr_accessor :destination_path
 
-      # Set the on-disk source file for this resource
+      # Get the on-disk source file for this resource
       # @return [String]
-      # attr_reader :source_file
-
+      # rubocop:disable AccessorMethodName
       def source_file
-        @source_file || get_source_file
+        if @source_file
+          @source_file
+        elsif proxy?
+          proxied_to_resource.source_file
+        else
+        end
       end
 
       # Initialize resource with parent store and URL
@@ -46,7 +53,7 @@ module Middleman
         @source_file = source_file
         @destination_path = @path
 
-        @local_metadata = { options: {}, locals: {} }
+        @metadata = Middleman::Sitemap::Extensions::Metadata.new(self)
       end
 
       # Whether this resource has a template file
@@ -54,26 +61,6 @@ module Middleman
       def template?
         return false if source_file.nil?
         !::Tilt[source_file].nil?
-      end
-
-      # Get the metadata for both the current source_file and the current path
-      # @return [Hash]
-      def metadata
-        result = store.metadata_for_path(path).dup
-
-        file_meta = store.metadata_for_file(source_file).dup
-        result.deep_merge!(file_meta)
-
-        local_meta = @local_metadata.dup
-        result.deep_merge!(local_meta)
-
-        result
-      end
-
-      # Merge in new metadata specific to this resource.
-      # @param [Hash] meta A metadata block like provides_metadata_for_path takes
-      def add_metadata(meta={})
-        @local_metadata.deep_merge!(meta.dup)
       end
 
       # The output/preview URL for this resource
@@ -90,6 +77,82 @@ module Middleman
         destination_path
       end
 
+      # Whether the Resource is ignored
+      # @return [Boolean]
+      def ignored?
+        if !proxy? && raw_data[:ignored] == true
+          true
+        else
+          @app.sitemap.ignored?(path) ||
+          (!proxy? &&
+            @app.sitemap.ignored?(source_file.sub("#{@app.source_dir}/", ''))
+          )
+        end
+      end
+
+      def render_options
+        metadata.fetch(:options)
+      end
+
+      def render_locals
+        metadata.fetch(:locals)
+      end
+
+      def add_metadata(meta)
+        metadata.add(meta)
+      end
+
+      # The preferred MIME content type for this resource
+      # Look up mime type based on extension
+      def content_type
+        mime_type = raw_data[:content_type] || render_options[:content_type] || ::Rack::Mime.mime_type(ext, nil)
+
+        return mime_type if mime_type
+
+        if proxy?
+          proxied_to_resource.content_type
+        else
+          nil
+        end
+      end
+
+      # Whether this page is a proxy
+      # @return [Boolean]
+      # rubocop:disable TrivialAccessors
+      def proxy?
+        @proxied_to
+      end
+
+      # Set this page to proxy to a target path
+      # @param [String] target
+      # @return [void]
+      def proxy_to(target)
+        target = ::Middleman::Util.normalize_path(target)
+        raise "You can't proxy #{path} to itself!" if target == path
+        @proxied_to = target
+      end
+
+      # The path of the page this page is proxied to, or nil if it's not proxied.
+      # @return [String]
+      attr_reader :proxied_to
+
+      # The resource for the page this page is proxied to. Throws an exception
+      # if there is no resource.
+      # @return [Sitemap::Resource]
+      def proxied_to_resource
+        proxy_resource = store.find_resource_by_path(proxied_to)
+
+        unless proxy_resource
+          raise "Path #{path} proxies to unknown file #{proxied_to}:#{store.resources.map(&:path)}"
+        end
+
+        if proxy_resource.proxy?
+          raise "You can't proxy #{path} to #{proxied_to} which is itself a proxy."
+        end
+
+        proxy_resource
+      end
+
       # Render this resource
       # @return [String]
       def render(opts={}, locs={})
@@ -97,10 +160,9 @@ module Middleman
 
         relative_source = Pathname(source_file).relative_path_from(Pathname(app.root))
 
-        instrument 'render.resource', path: relative_source, destination_path: destination_path  do
-          md   = metadata.dup
-          opts = md[:options].deep_merge(opts)
-          locs = md[:locals].deep_merge(locs)
+        instrument 'render.resource', path: relative_source, destination_path: destination_path do
+          opts = render_options.deep_merge(opts)
+          locs = render_locals.deep_merge(locs)
           locs[:current_path] ||= destination_path
 
           # Certain output file types don't use layouts
@@ -129,7 +191,7 @@ module Middleman
       #
       # @return [Boolean]
       def binary?
-        ::Middleman::Util.binary?(source_file)
+        source_file && ::Middleman::Util.binary?(source_file)
       end
     end
   end